Fix GCC warning

+++ src/tests/unit-tests/control-plane/gnss_block_factory_test.cc
@@ -285,7 +285,7 @@ TEST(GNSSBlockFactoryTest, InstantiateChannels)
     gr::msg_queue::sptr queue = gr::msg_queue::make(0);
     std::unique_ptr<GNSSBlockFactory> factory;
     std::unique_ptr<std::vector<std::unique_ptr<GNSSBlockInterface>>> channels = std::move(factory->GetChannels(configuration, queue));
-    EXPECT_EQ(2U, channels->size());
+    EXPECT_EQ(static_cast<unsigned int>(2), channels->size());
     channels->erase(channels->begin(), channels->end());
 }
